feat: ssr 방식 변경
